# zebra: Batch netlink messages