@eqvinox 2018-08-17
Hmm. "Stupid" question: does this actually help with anything?

The kernel processes netlink requests synchronously anyways; it's not like you can go do further processing while the kernel updates the routes... from what I remember it's the fastest to immediately read back the result too.

(or has this changed in the kernel?)