@vjardin 2018-08-18
It should increase the throughput of system calls (write). send*m*msg() shall be combined with it too.

Please confirm the benefits like any increase in the number of message rate.