@donaldsharp 2018-08-18
I was told by a kernel developer to do this because he felt it would be faster giving some of the internal mechanics of how routes are processed in the kernel.  But Vincent is correct we greatly reduce the task switching done.