@donaldsharp 2018-08-18
I tested a full bgp feed installation without this patch and with this patch on an ancient linux box.

Before:
annie# show thread cpu
Thread statistics for zebra:

Showing statistics for pthread main
-----------------------------------
                      CPU (user+system): Real (wall-clock):
Active   Runtime(ms)   Invoked Avg uSec Max uSecs Avg uSec Max uSecs  Type  Thread
    1          4.000         3     1333      4000      864      1452 R     zserv_accept
    0          0.000         1        0         0       43        43    E  frr_config_read_in
    0          0.000         3        0         0      215       270   T   if_zebra_speed_update
    0     134288.000     10903    12316    152000    12124    213241   T   work_queue_run
    1          0.000         3        0         0       64        85 R     vtysh_accept
    0       7460.000    116976       63    376000       63    374978    E  zserv_process_messages
    0          0.000         1        0         0      391       391   T   zebra_route_map_update_timer
    1        276.000        49     5632    268000     5215    247621 R     vtysh_read
    1          0.000         6        0         0       22        57 R     kernel_read


Showing statistics for pthread Zebra API client thread
------------------------------------------------------
                      CPU (user+system): Real (wall-clock):
Active   Runtime(ms)   Invoked Avg uSec Max uSecs Avg uSec Max uSecs  Type  Thread
    0          0.000        13        0         0       22        44  W    zserv_write
    1          4.000         9      444      4000       47       174 R     zserv_read


Showing statistics for pthread Zebra API client thread
------------------------------------------------------
                      CPU (user+system): Real (wall-clock):
Active   Runtime(ms)   Invoked Avg uSec Max uSecs Avg uSec Max uSecs  Type  Thread
    0          4.000        17      235      4000      153      2303  W    zserv_write
    1       4740.000    116966       40     12000       40     12131 R     zserv_read


Showing statistics for pthread Zebra API client thread
------------------------------------------------------
                      CPU (user+system): Real (wall-clock):
Active   Runtime(ms)   Invoked Avg uSec Max uSecs Avg uSec Max uSecs  Type  Thread
    0          0.000         1        0         0       29        29  W    zserv_write
    1          0.000         1        0         0       99        99 R     zserv_read


Total thread statistics
-------------------------
                      CPU (user+system): Real (wall-clock):
Active   Runtime(ms)   Invoked Avg uSec Max uSecs Avg uSec Max uSecs  Type  Thread
    7     146776.000    244952      599    376000      590    374978 RWTEX TOTAL

After:

annie# show thread cpu
Thread statistics for zebra:

Showing statistics for pthread main
-----------------------------------
                      CPU (user+system): Real (wall-clock):
Active   Runtime(ms)   Invoked Avg uSec Max uSecs Avg uSec Max uSecs  Type  Thread
    0          0.000         3        0         0      161       208   T   if_zebra_speed_update
    0      17204.000      1564    11000    128000    13330    195884   T   work_queue_run
    0          8.000         9      888      4000      592      1179   T   netlink_batch_expire
    1          0.000         5        0         0       67        84 R     vtysh_accept
    0          0.000         1        0         0       40        40    E  frr_config_read_in
    0          0.000         1        0         0      410       410   T   zebra_route_map_update_timer
    1       3636.000        70    51942    244000    52126    245487 R     vtysh_read
    0       7772.000    110420       70    388000       80    402951    E  zserv_process_messages
    1          4.000         3     1333      4000      855      1363 R     zserv_accept
    1          0.000        13        0         0       31        54 R     kernel_read


Showing statistics for pthread Zebra API client thread
------------------------------------------------------
                      CPU (user+system): Real (wall-clock):
Active   Runtime(ms)   Invoked Avg uSec Max uSecs Avg uSec Max uSecs  Type  Thread
    0          0.000        11        0         0       40       150  W    zserv_write
    1          0.000         7        0         0       45       184 R     zserv_read


Showing statistics for pthread Zebra API client thread
------------------------------------------------------
                      CPU (user+system): Real (wall-clock):
Active   Runtime(ms)   Invoked Avg uSec Max uSecs Avg uSec Max uSecs  Type  Thread
    0          0.000        14        0         0      224      2902  W    zserv_write
    1       4800.000    110412       43     16000       46     20086 R     zserv_read


Showing statistics for pthread Zebra API client thread
------------------------------------------------------
                      CPU (user+system): Real (wall-clock):
Active   Runtime(ms)   Invoked Avg uSec Max uSecs Avg uSec Max uSecs  Type  Thread
    0          0.000         1        0         0       30        30  W    zserv_write
    1          0.000         1        0         0      100       100 R     zserv_read


Total thread statistics
-------------------------
                      CPU (user+system): Real (wall-clock):
Active   Runtime(ms)   Invoked Avg uSec Max uSecs Avg uSec Max uSecs  Type  Thread
    7      33424.000    222535      150    388000      173    402951 RWTEX TOTAL

Testing methodology:

Start FRR, wait for convergence, then run a `show thread cpu`.

I am actually surprised by how much faster this was.  I will do more investigation, but this shows some promise.