@vjardin 2018-08-18
Still sendmsg() is used:
/* Send message to netlink interface. */
	frr_elevate_privs(&zserv_privs) {
		status = sendmsg(nl->sock, &msg, 0);
		save_errno = errno;
	}

instead of sendmmsg() http://man7.org/linux/man-pages/man2/sendmmsg.2.html

Since we get some batches now, sendmmsg() would add even more throughput.