@eqvinox 2018-08-21
Feedback on call:
- tunable timeout, because useful values can differ quite a bit between use cases
- might be possible to do something like SPF backoff, i.e. only start buffering after one or more requests have been sent unbatched
- definitely should have a well-chosen max number of batched messages to hit the "sweet spot", since after a certain number of batched messages the gains will diminish away