@donaldsharp 2018-08-23
I need to do the work outlined by @eqvinox above as well as add 1 more bit of code to better handle install failures in a asynchronous manner.